extract test for circular assembly references into dedicated test
